refactor(go): move query specs and validation function to a separate package
