+++ go/internal/apivalidator/types.go
@@ -1,0 +1,23 @@
+package apivalidator
+
+type SortOrder string
+
+const (
+	SortOrderAscendant  SortOrder = "ASC"
+	SortOrderDescendant SortOrder = "DESC"
+)
+
+type Criterion struct {
+	OperandLeft  string
+	OperandRight *string
+	Operator     string
+}
+
+type QueryInput struct {
+	Filter           *string      `json:"filter,omitempty"`
+	FilterExpression *[]Criterion `json:"filterExpression,omitempty"`
+	Limit            *int64       `json:"limit,omitempty"`
+	Offset           *int64       `json:"offset,omitempty"`
+	SortField        *string      `json:"sortField,omitempty"`
+	SortOrder        *SortOrder   `json:"sortOrder,omitempty"`
+}
