+++ go/service/policies/api_op_ListPolicies_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Think-iT-Labs/edc-connector-client/go/edc"
 	edchttp "github.com/Think-iT-Labs/edc-connector-client/go/edc/transport/http"
 	"github.com/stretchr/testify/assert"
+	"github.com/Think-iT-Labs/edc-connector-client/go/internal/apivalidator"
 )
 
 func Test_ListPolicies(t *testing.T) {
@@ -38,9 +39,9 @@ func Test_ListPolicies(t *testing.T) {
 	apiClient, err := New(*cfg)
 	assert.NoError(t, err, "failed to initialize api client")
 
-	policies, apiError, err := apiClient.ListPolicies(ListPoliciesInput{})
+	policies, apiError, err := apiClient.ListPolicies(apivalidator.QueryInput{})
 
-	assert.NoError(t, err, "failed to create asset.")
+	assert.NoError(t, err, "failed to list policies.")
 	assert.NotNil(t, policies)
 	assert.Nil(t, apiError)
 	assert.Equal(t, len(policies), 1)
@@ -77,7 +78,7 @@ func Test_ListPoliciesInternalServerError(t *testing.T) {
 	apiClient, err := New(*cfg)
 	assert.NoError(t, err, "failed to initialize api client")
 
-	policies, apiError, err := apiClient.ListPolicies(ListPoliciesInput{})
+	policies, apiError, err := apiClient.ListPolicies(apivalidator.QueryInput{})
 
 	assert.NoError(t, err, "failed to list policies.")
 	assert.Nil(t, policies)
